Some small fixes for RADIO, added support for login for members.

+++ db/migrate/20150327210544_create_ideal_transactions.rb
@@ -1,6 +1,6 @@
 class CreateIdealTransactions < ActiveRecord::Migration
   def change
-    drop_table :ideal_transactions
+    drop_table :ideal_transactions if ActiveRecord::Base.connection.table_exists? 'ideal_transactions'
     
     create_table :ideal_transactions, :id => false do |t|  
       t.string :uuid, :unique => true, :limit => 16
@@ -12,5 +12,8 @@ class CreateIdealTransactions < ActiveRecord::Migration
     
       t.timestamps
     end
+    
+    #create unique email for login
+    add_index :members, :email, :unique => true
   end
 end
